Added PhysicsModel class for use in PhysicsComponents. SAT-collision are being detected true/false only and are not handled, but they work :)

+++ src/com/detour/raw/GameManager.java
@@ -36,7 +36,7 @@ public class GameManager {
 	public void update(){
 		if(levelLoaded){
 			mEntityManager.update();
-			camera.update(hero.getX(), hero.getY());
+			camera.update((float)hero.getX(), (float)hero.getY());
 			mHUD.update();
 		}
 	}
